sauces and user routes improved with userId token specifications & multer code content improved with fileFilter properties

+++ controllers/sauces.js
@@ -6,39 +6,47 @@ exports.createSauce = (req, res, next) => {
     const sauceObject = JSON.parse(req.body.sauce);
     const sauce = new Sauce({
      ...sauceObject,
-     imageUrl: `${req.protocol}://${req.get('host')}/images/${req.file.filename}`,
-     //add likes dislikes usersliked usersdisliked here
-     /*likes: 0,
-        dislikes: 0,
-        usersLiked: [' '],
-        usersdisLiked: [' '],*/
+     imageUrl: `${req.protocol}://${req.get('host')}/images/${req.file.filename}`
     });
     sauce.save()
       .then(() => res.status(201).json({ message: 'Votre sauce est enregistrée !'}))
-      .catch((error) => res.status(400).json({ error }));
-  };
+      .catch(error => res.status(400).json({ error }));
+    };
 
 exports.modifySauce = (req, res, next) => {
+   Sauce.findOne({ _id: req.params.id })
+   .then(sauce =>{
     const sauceObject = req.file ?
     { 
       ...JSON.parse(req.body.sauce),
       imageUrl: `${req.protocol}://${req.get('host')}/images/${req.file.filename}`
     } : {...req.body};
 
-    Sauce.updateOne({ _id: req.params.id }, {...sauceObject, _id: req.params.id})
+    if(sauceObject.userId === sauce.userId) {
+      Sauce.updateOne({ _id: req.params.id }, {...sauceObject, _id: req.params.id})
       .then(() => res.status(200).json({ message: 'Votre sauce est modifiée !'}))
-      .catch((error) => res.status(400).json({ error }));
+      .catch(error => res.status(400).json({ error }));
+    } else {
+      res.status(401).json({ error: "Vous ne pouvez pas modifier cette sauce"})
+    }
+   }) 
+    .catch(error => res.status(500).json({ error }));  
   };
 
 exports.deleteSauce = (req, res, next) => {
-    Sauce.findOne({_id: req.params.id })
+    Sauce.findOne({ _id: req.params.id })
     .then(sauce => {
-      const filename = sauce.imageUrl.split('/images/')[1];
-      fs.unlink(`images/${filename}`, () => {
-        Sauce.deleteOne({ _id: req.params.id })
-          .then(() => res.status(200).json({ message: 'Votre sauce est supprimée!'}))
-          .catch((error) => res.status(400).json({ error }));
-      });
+      if (sauce.userId === req.token.userId) {
+        
+        const filename = sauce.imageUrl.split('/images/')[1];
+        fs.unlink(`images/${filename}`, () => {
+          Sauce.deleteOne({ _id: req.params.id })
+            .then(() => res.status(200).json({ message: 'Votre sauce est supprimée!'}))
+            .catch((error) => res.status(400).json({ error }));
+          });
+      } else {
+        res.status(401).json({ error: "Vous ne pouvez pas supprimer cette sauce"})
+      }
     })
     .catch((error) => res.status(500).json({ error }));
   };
